+++ src/main/java/net/minecraft/world/chunk/Chunk.java
@@ -787,7 +787,7 @@ public class Chunk {
     }
 
     public Random getRandomWithSeed(long seed) {
-        return new Random(this.worldObj.getSeed() + (this.xPosition * this.xPosition * 4987142) + (this.xPosition * 5947611) + (this.zPosition * this.zPosition) * 4392871L + (this.zPosition * 389711) ^ seed);
+        return new Random(this.worldObj.getSeed() + ((long) this.xPosition * this.xPosition * 4987142) + (this.xPosition * 5947611L) + ((long) this.zPosition * this.zPosition) * 4392871L + (this.zPosition * 389711L) ^ seed);
     }
 
     public boolean isEmpty() {
